+++ Classes/VTAcknowledgementsViewController.h
@@ -26,7 +26,7 @@
 @interface VTAcknowledgementsViewController : UITableViewController
 
 /**
- Custom font for the license main text view. If `nil` (default), the system font (size 16) is used.
+ Custom font for the license main text view. If `nil` (default), the system font (size 17) is used.
  */
 @property (nonatomic, strong) UIFont *licenseTextViewFont;