+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v0.2
 
-- Customizable font for license main text view
+- Non-editable license text view (duh)
+- Padding for license text view (iOS 7+ only)
+- Customizable font for license text view
 
 ## v0.1 (24 December 2013)